working on the settings menu

+++ settings.h
@@ -13,13 +13,13 @@ int backButtonX;
 int backButtonY;
 int thumbnailWidth;
 int thumbnailHeight;
-int mountainThumbX;
-int mountainThumbY;
-int mountainsThumbX;
-int mountainsThumbY;
-int japanThumbX;
-int japanThumbY;
-int spaceThumbX;
-int spaceThumbY;
+double mountainThumbX;
+double mountainThumbY;
+double mountainsThumbX;
+double mountainsThumbY;
+double japanThumbX;
+double japanThumbY;
+double spaceThumbX;
+double spaceThumbY;
 
 #endif //MY_AWESOME_PROJECT_SETTINGS_H
